Add linters and fixers to development dependencies
The linters and fixers are typically controlled by [pre-commit] so they
haven't been included in the `pyproject.toml` configuration. The down
side to this, however, is that any tooling that requires access to these
utilities can't find them. This is also the case with [mypy] and
[pytest].

By adding all of these libraries as development dependencies, though,
the developer experience can be improved. It also means that mypy can
find pytest's types so it no longer needs to be ignored.

[mypy]: http://www.mypy-lang.org
[pre-commit]: https://pre-commit.com
[pytest]: https://docs.pytest.org